+++ doctor/utils.py
@@ -1,4 +1,5 @@
-from django.utils.datastructures import SortedDict
+from collections import OrderedDict
+
 from django.views.debug import cleanse_setting
 
 
@@ -7,7 +8,7 @@ def cleanse_dictionary(dictionary):
     Cleanse sensitive values in a dictionary.
     """
 
-    cleansed_dictionary = SortedDict()
+    cleansed_dictionary = OrderedDict()
 
     for key, val in dictionary.iteritems():
         cleansed_dictionary[key] = cleanse_setting(key, val)